kern: Add support for SSP.
The kernel can now be built with stack smash protection.
It is by default enabled in the Debug build configuration.
Stack smash detection makes the compiler place stack canaries in
suspicous functions. At the end of the function the compiler will
check whether the canary has been overwritten. If this is the case it
will call __stack_chk_fail, allowing common memory handling errors on
the stack to be caught.

* kern/CMakeLists.txt: Add ssp subdirectory.
* kern/ssp: New kernel module.