romfs: Add comment about incorrect name assumption.
The current code is correct for traversing the directory structure,
however, there was a critical misunderstanding on my side: Names in
the file system are 16 byte *aligned*, not 16 byte in
total. Therefore, names can be arbitrarily long, which breaks our
code.
Even worse, this means that the API has to be changed to accomodate
variably sized strings.

* lib/romfs/include/romfs.h: Document wrong assumptions in API.
* lib/romfs/romfs.c: Document places that need fixing for long
  filenames.