boot: Initial successful boot!
With this version of the source code the kernel is correctly fired up
and booting. While the kernel still ignores the memory info passed to
it, and the bootloader does not fully construct it, this is quite a
success. The GOT relocations work correctly.
There's a bunch of junk in the repo right now that I need to weed out
though.