Fix hardfault after enabling interrupts in flash kernel.
The kernel had a problem where, when it was loaded from flash, it
crashed right after enabling interrupts.
The processor would try to enter a hardfault exception, which failed,
due to a malformed VTOR register entry. The reason was, that the
RP2040 has the first 8 bits of the VTOR register implemented as
RAZ/WI. This is allowed by the ARMv6m architecture reference manual,
but was not respected by the current code.
In conjunction with the lacking debugging support for flash kernels
before the newly created script, I incorrectly diagnosed that the VTOR
entries would not have been relocated, and have translated the vector
translation unit back to an assembler file. This created a new list of
problems. After the original problem is resolved, the vector file is
reverted back to the old C version and amended.

TLDR; This commit gets interrupts working correctly in both versions
of the kernel, but still has a crashing kernel, due to a bug in the
kalarm system.

* boot/boot.c (__vector): Renamed to __boot_vector.
  It is really just a crutch that we need for the boot2 payload to
  jump to our bootloader. However, it does not handle interrupts.
  Naming the variable __vector does active harm, as it shadows the
  same variable in the kernel when debugging with GDB.
  Therefore, rename to a stub variable name that is unique.
* kern/vector.s: Deleted.
* kern/vector.c: Resurrected old version.
  Furthermore, the __vector variable is placed in its own subsection
  of the .data section. Despite it being constant, it needs to be
  allocated in SRAM as its contents need relocation before execution
  of the kernel.
  Finally, adapt the alignment of the variable to what the RP2040
  manual requires.
* kern/CMakeLists.txt: Replace vector.s back to vector.c
* lib/elf/elf-sniffer.gdb: Rename to .py extension.
  GDB doesn't understand that the content is python code otherwise.
* lib/elf/elf.c (memset): New, weak function.
  Allows for more fluent stepping through the load_elf_file function
  and might profit from a faster, strongly defined memset version.
  (load_elf_file): Use new memset function, instead of a raw for loop.