kern: Use libheap and fix problem of hard-faulting ISR.
The kernel crashed in the systick ISR, because GCC generated
completely bogus code to reference the heap data structure with
PC-relative accesses. Turns out, GCC assumed this to be possible in
our compilation profile and I was just lucky so far. Fixed by
explicitly forbidding GCC to access data PC-relative.

* kern/CMakeLists.txt: Link libheap.
* kern/kalarm.c: Switch to using heap from libheap.
* kern/include/kern/debug.h: Add kassert macro.
* kern/platform/rp2040/CMakeLists.txt: Compile with
  `-mno-pic-data-is-text-relative`.