kern: Restructure kalarm.
There was a bug in the kernel, where syscalls, leading to reschedules
of threads would flood the kalarm heap. There must be only reschedule
event in the kalarm heap at all times, as each event execution will
schedule a new one.
Restructure the kalarm system to allow to update event types instead
of queing new events of the same type.

* kern/include/kern/kalarm.h: Update declarations.
* kern/kalarm.c (kalarm_event): Use new enum for definition.
  (find_first_kalarm_event): New function.
  (register_kalarm_event): Redefine according to new interface.
  (update_kalarm_event): New function.
* kern/systick.c (request_next_thread): New helper function.
  (kalarm_event_funcs): New lookup table.
  (systick_init): Reduce tick rate.
  (schedule_next_thread): Only update next thread instead of queing a
  new one.