kern: Fix critical bug in kalarm system.
The debug build of the operating system would consistently hard fault
when switching contexts. This was because the invoked pendsv handler
would corrupt the kalarm heap by means of a bug in
update_kalarm_event, after which the systick handler would pick up an
almost-nullptr and jump to it, causing the hardfault. I guess this
was only apparent in debug mode as the interrupt timing is slightly
different.

* kern/kalarm.c (register_kalarm_event): Explictily zero missing
  struct member.
  (update_kalarm_event): Fix swapped arguments.