kern: Move syscall_thread_control to new translation unit.
The thread control logic takes quite some lines of source code and is
pretty tied into the internal logic of thread.c. Move it to its own
translation unit to improve the encapsulation of thread.c and have the
logic separated.
Furthermore, this is the first step to get all the syscalls into
separate translation units.

* kern/CMakeLists.txt: Add new translation unit.
* kern/include/kern/syscall.h: New file.
* kern/include/kern/thread.h (get_tcb_index): New declaration.
  (delete_thread): New declaration.
  (write_utcb): New declaration.
* kern/thread.c (delete_thread): New function.
  (write_utcb): Mark as non-static.
  (get_tcb_index): New function.
  (syscall_thread_control and friends): Deleted.
* kern/syscall.c (softirq_svc): Delete external declarations and rely
  on header includes.
* kern/syscall/thread_control.c: New file.