kern: Restructure IPC syscall to keep processing partner IPC.
If we find a sending thread for our receive IPC, it is important that
we process the following receive phase in the partner thread as well,
to keep the system from stalling.
This requires some restructuring as the code cannot assume anymore
that the target of the current operation is stored in `caller`.