[WIP] vfs: Add VFS_MOVE operation.
We cannot fully implement something like POSIX dup, as it would
require the server to be able to share state between fds.
Instead, I'm adding move for now, which allows the same file to get a
different handle assigned.
This is required in the bsh in order to do directory traversal
properly.
With this new feature the shell builtins should finally be properly
implementable.

* lib/vfs/include/vfs.h (enum vfs_ops): Add VFS_MOVE.
  (enum vfs_move_args): New enum.
  (enum vfs_answer_ops): Add VFS_MOVE_RET.
  (enum vfs_move_ret_args): New enum.
* lib/vfs/include/vfs/client.h (readdir): Move into declaration order.
  (move_fd): New function declaration.
* lib/vfs/vfs-client.c (readdir): Move into declaration order.
  (move_fd): New function.
* usr/root/romfs_server.c (handle_vfs_move): New function.
  (romfs_main): Use handle_vfs_move to handle new VFS_MOVE request.