pl011: De-float baud rate calculation.
There's a bug in the toolchain at the moment where in the debug build
the link step will fail when doing double divisions because of
relocations in read-only segments.
This is not even mitigated when recompiling libgcc with split-ELF
flags.
In order to properly solve this issue, more investigation is
required. For now, the logic is simply restated to not use any
floating point math.

* usr/pl011/pl011.c (uart_set_baud): Adjust logic to not use float
  math.
  Take a new parameter for the upstream clock's frequency.
  (uart_init): Adjust for new parameters of uart_set_baud.