lib/rp2040_utils: New library.
Source out some general rp2040 code from the root executable into a
library. Especially subsystem resets will be performed in drivers
until there's a proper service for them.
In general it makes sense to have the hardware interface be formulated
a bit more generically and less in the context of the root process.

* lib/rp2040_utils: New library.
* lib/CMakeLists.txt: Add rp2040_utils to subdirectories.
* lib/root/CMakeLists.txt: Adapt for some source files moving to new
  library.
* lib/root/clocks.c: Adapt include paths.