kern: Fix kernel panic on thread switching.
The kernel was crashing semi-regularly because of the panic in
`find_thread_idx_to_schedule_from_target()`:
```
panic("Thread %u is not active, but shall be scheduled\n",
      schedule_target->global_id);
```
This happened because, `schedule_target` was changed to `NULL` while
the interrupt was executed. This can happen, as pendsv is the
interrupt with the lowest priority.
Mitigate this for now by disabling interrupts slightly earlier.

* kern/thread.c (schedule_next_thread): Disable interrupts before
  reading `schedule_target`.